Ant Game playable. needs finishing touches
There is now dynamic compilation and reflection of all user code into
the application.

However a lot of downfalls with the application and with the default
approach to programming the ants have been shown as well. For example
the inability to define variables, or custom states for the ants to be
in hinder the progress of the students. This curls back to a core
point. When do you start to introduce other power enabling ideas such
as types, iteration, syntax, OO and classes, where users would then be
able to implement their idea of the ideal ant?

As far as the application code goes. There are a good few helper
methods still needed un order to make the basic functions clear for the
average user, as well as some advanced suggestions or guides for the
more advanced who would like to use java to the full.

+++ src/me/paulrose/lptc/simulator/ColonyFactory.java
@@ -5,7 +5,7 @@ public class ColonyFactory extends BaseFactory
 
 	ColonyFactory(World w){ super(w); }
 
-	public Colony create(String n, int x, int y, int f, String atype)
+	public Colony create(String n, float x, float y, int f, String atype)
 	{		
 		Colony c = new Colony(n, x, y, atype, 100, world);
 		world.addEntity(c);
